Refactor test code to main.py, graph chunks individually to verify correctness
